backend/pyproject.toml
[project]
name = "trpg-world-backend"
version = "0.1.0"
description = "TRPG World Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "python-socketio>=5.10.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "litellm>=1.0.0",
    "langgraph>=0.0.1",
    "langchain-core>=0.1.0",
    "langchain-community>=0.4.1",
    "langchain-litellm>=0.3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.92.0",
    "httpx>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiohttp>=3.13.2",
    "ruff>=0.14.10",
]
